fix del img

+++ handlers/product_handler.go
@@ -472,37 +472,39 @@ func DeleteProduct(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err})
 		return
 	}
-	// delete foundProduct.Image
-	client, err := storage.NewClient(ctx, option.WithCredentialsFile("ezcom-eaa21-firebase-adminsdk-9zpt0-d8e4765278.json"))
-	if err != nil {
-		log.Fatalf("Failed to create client: %v", err)
+	if foundProduct.Image != "" {
+		// delete foundProduct.Image
+		client, err := storage.NewClient(ctx, option.WithCredentialsFile("ezcom-eaa21-firebase-adminsdk-9zpt0-d8e4765278.json"))
+		if err != nil {
+			log.Fatalf("Failed to create client: %v", err)
+		}
+		defer client.Close()
+
+		// ชื่อของ bucket ที่เก็บไฟล์
+		bucketName := "ezcom-eaa21.appspot.com"
+
+		// ชื่อของไฟล์ที่ต้องการลบ
+
+		// Example string
+		path := foundProduct.Image
+
+		// Split the string using "/"
+		parts := strings.Split(path, "/")
+
+		// Print the last element
+		lastIndex := len(parts) - 1
+		parts1 := strings.Split(parts[lastIndex], "?")
+		fmt.Println(parts1[0])
+		fileName := parts1[0]
+
+		// ลบไฟล์
+		err = client.Bucket(bucketName).Object(fileName).Delete(ctx)
+		if err != nil {
+			log.Fatalf("Failed to delete object: %v", err)
+		}
+
+		fmt.Println("Object deleted successfully")
 	}
-	defer client.Close()
-
-	// ชื่อของ bucket ที่เก็บไฟล์
-	bucketName := "ezcom-eaa21.appspot.com"
-
-	// ชื่อของไฟล์ที่ต้องการลบ
-
-	// Example string
-	path := foundProduct.Image
-
-	// Split the string using "/"
-	parts := strings.Split(path, "/")
-
-	// Print the last element
-	lastIndex := len(parts) - 1
-	parts1 := strings.Split(parts[lastIndex], "?")
-	fmt.Println(parts1[0])
-	fileName := parts1[0]
-
-	// ลบไฟล์
-	err = client.Bucket(bucketName).Object(fileName).Delete(ctx)
-	if err != nil {
-		log.Fatalf("Failed to delete object: %v", err)
-	}
-
-	fmt.Println("Object deleted successfully")
 
 	var collection = db.GetSpecs_Collection()
 	objSpecsID, err := primitive.ObjectIDFromHex(foundProduct.Specs)
